refactor: use modern NodeJS code
